ml-analytics/requirements.txt
paho-mqtt==1.6.1
prometheus-client==0.19.0
pandas==2.2.2
scikit-learn==1.5.1
numpy==2.0.1
